+++ src/Faker/Extensions/IEnumerableExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace RimuTec.Faker.Extensions {
+   internal static class IEnumerableExtensions {
+      public static IOrderedEnumerable<T> Shuffle<T>(this IEnumerable<T> source) {
+         // Ref: Comment by user 'grenade' on answer https://stackoverflow.com/a/4262134/411428
+         return source.OrderBy(x => RandomNumber.Next());
+      }
+   }
+}